requirements.txt
torch>=2.1
transformers>=4.36
pydantic>=2.5
fastapi>=0.104
uvicorn>=0.24
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.66
mlflow>=2.9
google-genai>=1.0
matplotlib>=3.8
seaborn>=0.13
pytest>=7.4
ruff>=0.1
httpx>=0.25
scipy>=1.11
